base-files: add support for hex root parameter
For some devices, such as the Turris Omnia running TurrisOS 4, the root
parameter is neither using a /dev/<disk_name> representation, nor a
PARTUUID, but a major/minor number in hexadecimal format such as:
"b301".

On those device, the previous export_bootdevice() function would fail
to identify the boot device, and sysupgrade would error out with:
"Unable to determine upgrade device"

This commit adds a case for the <hex_major><hex_minor> format. The
major and minor of the partition are first stored in the MAJOR and
MINOR variables respectively, which will later be overwritten with the
major and minor of the boot device when reading the uevent file.

Note that the "<hex_major>:<hex_minor>" format is also allowed by the
kernel and not handled here, but support for it could easily be added
when/if it's needed.

For reference, the supported root parameter values are documented in
in the kernel in the "name_to_dev_t" comment in init/do_mounts.c.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>